+++ src/steps/stepInterfaces.ts
@@ -4,7 +4,7 @@ import { AliasedTable, Table } from '../database'
 import { PrimitiveType } from '../models/types'
 import { Condition } from '../models/Condition'
 import { OnStep } from './OnStep'
-import { WhereStep } from './WhereStep'
+import { SelectWhereStep } from './SelectWhereStep'
 import { Column } from '../columns'
 import { OrderByArgsElement } from '../orderBy'
 import { All } from '../singletoneConstants'
@@ -43,11 +43,11 @@ export interface IAfterFromSteps extends BaseStep, OrderByStep {
 
   fullOuterJoin(table: Table): FullOuterJoinStep
 
-  where(condition: Condition): WhereStep
+  where(condition: Condition): SelectWhereStep
 
-  where(left: Condition, operator: LogicalOperator, right: Condition): WhereStep
+  where(left: Condition, operator: LogicalOperator, right: Condition): SelectWhereStep
 
-  where(left: Condition, operator1: LogicalOperator, middle: Condition, operator2: LogicalOperator, right: Condition): WhereStep
+  where(left: Condition, operator1: LogicalOperator, middle: Condition, operator2: LogicalOperator, right: Condition): SelectWhereStep
 
   groupBy(...groupByItems: Column[]): GroupByStep
 
@@ -81,9 +81,9 @@ export interface OnOrStep extends OnStep {}
 
 export interface OnAndStep extends OnStep {}
 
-export interface WhereOrStep extends WhereStep {}
+export interface WhereOrStep extends SelectWhereStep {}
 
-export interface WhereAndStep extends WhereStep {}
+export interface WhereAndStep extends SelectWhereStep {}
 
 export interface GroupByStep extends BaseStep, OrderByStep {
   having(condition: Condition): HavingStep